Rydder litt før ny feature for automatisk revurdering. Refaktorering for å skille bedre mellom inntektsendring og vedtaksendring. Vedtaksendring logikk i egen util-klasse.
