+++ src/main/kotlin/no/nav/familie/ef/personhendelse/inntekt/AutomatiskRevurderingService.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.ef.personhendelse.inntekt
+
+import org.springframework.stereotype.Service
+
+@Service
+class AutomatiskRevurderingService {
+    fun harStabilInntekt() {
+    }
+}
