+++ src/main/kotlin/no/nav/familie/ef/personhendelse/inntekt/InntektController.kt
@@ -11,11 +11,11 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping(path = ["/api/inntekt"])
 @ProtectedWithClaims(issuer = "azuread")
 class InntektController(
-    val vedtakendringer: VedtakendringerService,
+    val inntektsendringerService: InntektsendringerService,
 ) {
     @GetMapping("/sjekkEndringer")
     fun sjekkEndringer(): ResponseEntity<Any> {
-        vedtakendringer.beregnInntektsendringerAsync()
+        inntektsendringerService.beregnInntektsendringerAsync()
         return ResponseEntity.ok().build()
     }
 
@@ -23,8 +23,8 @@ class InntektController(
     fun opprettOppgaverForInntektsendringer(
         @RequestParam skalOppretteOppgaver: Boolean,
     ): ResponseEntity<Int> {
-        val antallOppgaver = vedtakendringer.opprettOppgaverForInntektsendringer(skalOppretteOppgaver)
-        vedtakendringer.opprettOppgaverForNyeVedtakUføretrygd()
+        val antallOppgaver = inntektsendringerService.opprettOppgaverForInntektsendringer(skalOppretteOppgaver)
+        inntektsendringerService.opprettOppgaverForNyeVedtakUføretrygd()
         return ResponseEntity.ok(antallOppgaver)
     }
 }
